Refactor constants module to improve code cohesion
