+++ src/aws/mod.rs
@@ -5,6 +5,9 @@ pub mod credentials;
 pub mod roles;
 pub mod sts;
 
+/// Default AWS region for STS operations when no region is configured
+pub const DEFAULT_AWS_REGION: &str = "us-east-1";
+
 /// AWS temporary credentials structure
 #[derive(Debug, Clone)]
 pub struct Credentials {
