+++ src/lib.rs
@@ -3,6 +3,5 @@ pub mod browser;
 pub mod cli;
 pub mod commands;
 pub mod config;
-pub mod constants;
 pub mod idp;
 pub mod saml;
